save: make save files mostly compatible with Gens
Since Gens and Kega save file formats are also based on Genecyst, I
decided to make DGen/SDL compatible with them again thanks to the
documentation in Gens' source code.

DGen is now able to store its sound state properly. Old save files are
still supported with the usual sound glitches.

+++ md.h
@@ -131,7 +131,7 @@ typedef struct {
 	uint32_t d[8]; /* D0-D7 */
 	uint32_t a[8]; /* A0-A7 */
 	uint32_t pc;
-	uint32_t sr;
+	uint16_t sr;
 } m68k_state_t;
 
 typedef struct {
